+++ CHANGELOG.md
@@ -1,6 +1,15 @@
 # CHANGELOG
 All notable changes to this project will be documented in this file. Note: the first version tracked here is 2.0.0. Previous versions were before the project's migration to Vue and were not tracked.
 
+## [2.0.1] - 2021-10-09
+#### Added:
+
+  - Win percentage tooltip on hover to Teams vs. Opponents page
+
+#### Changed:
+
+  - Updated Teams vs. Opponents page to parse and show data from JSON file
+
 ## [2.0.0] - 2020-09-06
 #### Added: